cleanup and moving some stuff around
